adding delete confirmation

+++ src/main/kotlin/com/github/mckernant1/minecraft/jocky/commands/DestroyCommand.kt
@@ -4,8 +4,10 @@ import com.github.mckernant1.minecraft.jocky.cfn.cfnClient
 import com.github.mckernant1.minecraft.jocky.core.waitForCompletion
 import com.github.mckernant1.minecraft.jocky.execptions.InvalidCommandException
 import com.github.mckernant1.minecraft.jocky.singletons.serverTable
+import kotlinx.coroutines.delay
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent
 import software.amazon.awssdk.services.cloudformation.model.StackStatus
+import java.time.Duration
 
 class DestroyCommand(event: MessageReceivedEvent) : AbstractCommand(event) {
     override fun validate(): Unit {
@@ -14,6 +16,22 @@ class DestroyCommand(event: MessageReceivedEvent) : AbstractCommand(event) {
 
     override suspend fun execute() {
         val server = server!!
+
+        event.channel.sendMessage(
+            "You are requesting to delete the server. This requires 2 member confirmation\n" +
+                    "Another member of the server must type: `\$confirm delete ${server.serverName}`"
+        ).complete()
+        delay(Duration.ofSeconds(30).toMillis())
+
+        val confirmation = event.channel.history.retrievedHistory.any {
+            it.author.id != event.author.id && it.contentRaw == "\$confirm delete ${server.serverName}"
+        }
+
+        if (!confirmation) {
+            event.channel.sendMessage("You have failed to confirm deletion").complete()
+            return
+        }
+        
         serverTable.deleteItem(server)
         cfnClient.deleteStack {
             it.stackName(server.getStackName())
